refactor: Convert course stage comment model to typescript
